docker/docker_requirements.txt
cython==0.29.7
subprocess32==3.5.4
biopython==1.73
pandas==0.24.2
numpy==1.16.4
tzlocal==1.5.1
psutil==5.4.7
awscli==1.16.15